requirements.txt
# Furryville Index - Python Requirements
# Flask web framework and dependencies
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0

# Database connectivity (PyMySQL instead of mariadb for better Python 3.12 compatibility)
PyMySQL==1.1.1

# Environment variables management
python-dotenv==1.1.1

# System utilities (Windows compatibility)
colorama==0.4.6
packaging==25.0
